==> Phising_Sighting/1.Collecting-data.ipynb <==
#Importing libraries
import numpy as np
import pandas as pd


# Hugging Face dataset: https://huggingface.co/datasets/pirocheto/phishing-url
# 7000 instances

url = "https://huggingface.co/datasets/pirocheto/phishing-url/resolve/main/data/train.parquet"
hugf = pd.read_parquet(url)
hugf.head()
hugf.shape


#Hugging Face dataset : https://huggingface.co/datasets/bgspaditya/phishing-datas
# 600,000 instances


# Phish Tank dataset
# ISSUE: No features are present. 
phishtank = pd.read_csv("data/verified_online.csv")
phishtank = phishtank[["url" ]]
phishtank.head()


df.shape

phishes["url"] = df[["url"]]


import pandas as pd

url = "https://huggingface.co/datasets/pirocheto/phishing-url/resolve/main/data/train.parquet"
hugface = pd.read_parquet(url)
