==> Phising_Sighting/3.Large-datasets.ipynb <==
# Datasets available


## 1. Kaggle  
https://www.kaggle.com/datasets/harisudhan411/phishing-and-legitimate-urls/data

import pandas as pd
data0 = pd.read_csv("data/1.URL-Kaggle-800k.csv")
data0.shape

data0.head()

## 2. Hugface - bgspaditya  
https://huggingface.co/datasets/bgspaditya/phishing-dataset

data1 = pd.read_csv("data/2.hf-aditya.csv")
data1.shape

data1.head()

### 1. Merging this dataset with previous one

# merging Dataset0 and Dataset1
merge01 = pd.merge(data0, data1, left_on = ["url"], right_on = ["url"], how = "outer")
merge01 = merge01.reset_index()
merge01.head()

#What are all these random symbols, surely they are not actual urls.
merge01.shape
#This shows there are 602 new records that are not supposed to be there

#pd.merge(merge01, data0, on=['url'], how = "inner").shape
test = pd.merge(merge01, data1, on=['url'], how = "inner")



Issues with merging:  
1. Duplicate urls. 
2. Urls with contradicting labels. i.e 1 and benign. or 0 and phishing
3. data1 has values other than phishing and benign. - solved
4. urls 

### data1 has values other than phishing and benign - solved

#debugging
merge01["type"] == "defacement"

# Removing rows with type defacement, as that is not what we need
merge01 = merge01.drop(merge01[(merge01["status"].isna() ) & (merge01["type"] == "defacement")].index) 

# label malware data as phishing data 
merge01 = merge01.replace("malware", value="phishing")
merge01["type"].unique()

### creating label column - issue unsolved

# renaming columns
merge01.columns = ["index", "url", "A", "B"]
merge01.head()


# number of rows with contradicting labels
n_error = sum((merge01["A"] == 1) & (merge01["B"] == "phishing")) + sum((merge01["A"] == 0) & (merge01["B"] == "benign"))
n_total = len(merge01["index"])
per_error = n_error/n_total
print(per_error*100)

### Duplicates

merge01.nunique()

merge01 = merge01.drop_duplicates("url")
merge01.nunique()

data0.shape

data0.nunique()

data0.drop_duplicates().shape

data1.shape

data1.nunique()

data1.drop_duplicates().shape

## *MERG01* now has no duplicates, and each row contains a label from each dataset, A and B.

### Create new label "A" column switching labels to 0 for benign, 1 for phishing

merge01.head()

# Creating a new column with switched values
import numpy as np
merge01["A"] = merge01["A"].map({1:0 , 0:1})
merge01.head()

# making a new copy so i don't make dumb mistakes like i just did right now
mergeAB = merge01.copy(deep=True)

mergeAB.columns = ["index", "url", "label_A", "label_B"]
mergeAB.head()

### Split into "contradicting data" & labelled data  
1. 0 and Phish & 1 and benign

#### Get and save unlabelled data

labelled_url1 = mergeAB[((mergeAB["label_A"] == 1) & (mergeAB["label_B"] == "phishing")) | ((mergeAB["label_A"] == 0) & (mergeAB["label_B"] == "benign"))]
labelled_url2 = mergeAB[(mergeAB["label_A"].isna()) | (mergeAB["label_B"].isna())]
labelled_url = pd.merge(labelled_url1, labelled_url2, how = "outer")
labelled_url.shape

unlabelled_url = mergeAB[ (mergeAB["label_B"] == "phishing") & (mergeAB["label_A"] == 0) | (mergeAB["label_B"] == "benign") & (mergeAB["label_A"] == 1)]

unlabelled_url = unlabelled_url.drop(columns= ["label_A", "label_B"])

unlabelled_url = unlabelled_url.reset_index()
unlabelled_url.to_csv('URL_to_test.csv')

unlabelled_url.to_csv("URL_to_test.csv")
unlabelled_url.head()

#### Get labeleld data and clean up

print("Labelled url:", labelled_url.shape)
print("Unlabelled_url: ", unlabelled_url.shape)
print("mergeAB: ", mergeAB.shape)

labelled_url.head()

labelled_url["label_B"] = labelled_url["label_B"].map({"phishing":1 , "benign":0})
labelled_url.head()

labelled_url.reset_index(inplace=True)

labelled_url = labelled_url.drop(columns = ["index"])

labelled_url.loc[labelled_url['level_0'] == 0]

def labelling(row):
    if (row.label_A == 1) | (row.label_A == 0):
        return row.label_A
    else:
        return row.label_B
    
def labelling1(row):
    return row["label_A"]


labelled_url["label"] = labelled_url.apply(labelling, axis =1)

labelled_url = labelled_url.drop(columns=["label_A", "label_B"])
print(labelled_url.shape)

sum(labelled_url["label_A"].isna())

check = labelled_url[["label_A", "label_B"]].isna()
check[  (check["label_B"] == True)]

labelled_url

labelled_url = labelled_url.drop( columns= ["level_0"])
labelled_url.to_csv("Phishing_URL_0.csv")
