==> Phising_Sighting/4.Adding-more-data.ipynb <==
import pandas as pd
import numpy as np

data1 = pd.read_csv("data/3.Alexkstern.csv")
data1.head()

data1 = data1.drop(columns = ["Unnamed: 0"])
#data1.columns = ["index1", "url", "label1"]
data1.head()

data1.columns = ["url", "label"]
#data1.head()

data0 = pd.read_csv("Phishing_URL_0.csv")
data0.head()

data0 = data0.drop(columns = ["Unnamed: 0"])
#data0.columns = ["index0", "url", "label0"]
data0.head()

merge01 = pd.merge(left = data0, right = data1, on= "url", how="outer")
merge01.head()

#Dropping these not able to read encodings
merge01 = merge01.reset_index()


#print(sum(merge01.index == 10))
print(merge01.shape)
merge01 = merge01.drop([x for x in range(0,10)])
#print(sum(merge01.index == 10))
print(merge01.shape)
merge01.head()

merge01 = merge01.drop([x for x in range(909621,909668)])

merge01.reset_index(drop=True, inplace = True)
merge01.head()

merge01 = merge01.drop(["level_0", "index"], axis =1)
merge01.shape

merge01 = merge01.drop_duplicates("url")


merge01.shape

labelled_url1 = merge01[ merge01["label_x"] == merge01["label_y"] ]
labelled_url2 = merge01[ (merge01["label_x"].isna()) | (merge01["label_y"].isna())  ]
labelled_url = pd.merge(labelled_url1, labelled_url2, how = "outer")
labelled_url.head()

labelled_url.shape

def labelling(row):
    if (row.label_x == 1) | (row.label_x == 0):
        return row.label_x
    else:
        return row.label_y
    
def labelling1(row):
    return row["label_A"]


labelled_url["label"] = labelled_url.apply(labelling, axis =1)

labelled_url.head()

labelled_url.drop(["label_x", "label_y"], axis = 1, inplace=True)
labelled_url.head()

labelled
