==> Phising_Sighting/5. Adding-Mitake-dataset.ipynb <==
import pandas as pd
import numpy as np

data1 = pd.read_csv("data/4.Mitake.csv")
print(data1.shape)
data1.head()

data0 = pd.read_csv("Phishing_URL_1.csv")
data0.head()

data0 = data0.drop(columns = ["Unnamed: 0"])
#data0.columns = ["index0", "url", "label0"]
data0.head()

merge01 = pd.merge(data0,data1,on="url",how="outer")

merge01 = merge01.drop_duplicates("url")

merge01 = merge01.reset_index()
merge01.head(15)

merge01 = merge01.drop("index", axis=1)

merge01.head()

#1270136 1270088
#merge01 = merge01.drop([x for x in range(1270088,1270136)])
merge01 = merge01.reset_index()
merge01.tail()


merge01.shape

label1 = merge01[ merge01["label_x"] == merge01["label_y"] ]
label2 = merge01[ (merge01["label_x"].isna()) | (merge01["label_y"].isna())  ]
labelled = pd.merge(label1, label2, how = "outer")
labelled.head()

def labelling(row):
    if (row.label_x == 1) | (row.label_x == 0):
        return row.label_x
    else:
        return row.label_y
    
def labelling1(row):
    return row["label_A"]


labelled["label"] = labelled.apply(labelling, axis =1)
labelled.drop(["label_x", "label_y"], axis = 1, inplace=True)

labelled =labelled.drop(["level_0", "index"], axis=True)
labelled.head()

labelled.to_csv("Phishing_URL_2.csv")

sum(labelled["label"] == 1)/labelled.count()
