==> Phising_Sighting/6.Lastone.ipynb <==
import pandas as pd

train = pd.read_csv("data/train.csv")
test = pd.read_csv("data/test.csv")

test.head(20)

data1 = pd.concat([train,test])
data1.columns = ["url", "label"]
data1.head()

data0 =  pd.read_csv("Phishing_URL_2.csv")
data0 = data0.drop("Unnamed: 0", axis=1)
data0.head()

#Create a function to merge two tables with binary labels
def merge_labelled(data0, data1, column):

    merged = pd.merge(left= data0, right = data1, on = column, how = "outer")
    



merged = pd.merge(left= data0, right = data1, on = "url", how = "outer")
merged.head()

label1 = merged[ merged["label_x"] == merged["label_y"] ]
label2 = merged[ (merged["label_x"].isna()) | (merged["label_y"].isna())  ]
labelled = pd.merge(label1, label2, how = "outer")
labelled.head()

def labelling(row):
    if (row.label_x == 1) | (row.label_x == 0):
        return row.label_x
    else:
        return row.label_y
    
def labelling1(row):
    return row["label_A"]


labelled["label"] = labelled.apply(labelling, axis =1)
labelled.drop(["label_x", "label_y"], axis = 1, inplace=True)

print(labelled.head())
